requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.10
langchain-groq==0.0.1  # Note: using older version
langgraph==0.0.15
langsmith==0.0.92  # Pin the older version
duckduckgo-search==4.1.1
pypdf==3.17.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.1
pdfplumber==0.10.3
protobuf==4.25.8